requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
pydantic>=1.8.2
beautifulsoup4>=4.9.3
aiohttp>=3.7.4
Pillow>=8.3.1
python-multipart>=0.0.5
